Move propsFromAccountStatus function to AdSense util file and add case for connected account with zero data.
